+++ index.ts
@@ -1,2 +1,3 @@
 export { isPrime } from "./src/is_prime.js";
 export { integerSqrt } from "./src/integer_sqrt.js";
+export { gcd } from "./src/gcd.js";